fix: implement user/orders service test. [EVP-1774]

+++ src/main/kotlin/com/example/tutorial/dto/OrderCreateReq.kt
@@ -14,7 +14,7 @@ data class OrderCreateReq(
     val name: String?,
     @field:NotNull(message = "order price must not be null")
     @field:Max(value = 999999, message = "order price must not exceed 999,999")
-    val price: Long?,
+    val price: Double?,
     val address: String?,
     val orderDateTime: LocalDate?,
     val orderStatus: OrderStatus? = PENDING,
