+++ src/main/kotlin/com/example/tutorial/dto/OrderCreateRes.kt
@@ -7,7 +7,7 @@ import java.util.UUID
 data class OrderCreateRes(
     val id: UUID,
     val name: String? = null,
-    val price: Long? = null,
+    val price: Double? = null,
     val address: String? = null,
     val orderDateTime: LocalDate? = null,
     val orderStatus: OrderStatus,
